requirements.txt
Flask
flask_cors
paddleocr
pillow
numpy
tqdm
